+++ lax/displayer.h
@@ -212,7 +212,7 @@ class Displayer : public PanUser, virtual public anObject
 	virtual void drawthing(flatpoint p, double rx, double ry, int tofill, DrawThingTypes thing);
 	virtual void drawthing(double x, double y, double rx, double ry, int tofill, DrawThingTypes thing); // draws same orientation on screen
 	virtual void drawthing(double x, double y, double rx, double ry, DrawThingTypes thing,unsigned long fg,unsigned long bg,int lwidth=1);
-	virtual void drawarrow(flatpoint p,flatpoint v,int rfromp=0,double len=10,char reallength=1,int portion=3,bool center=false);
+	virtual void drawarrow(flatpoint p,flatpoint v,double rfromp=0,double len=10,char reallength=1,int portion=3,bool center=false);
 	virtual void drawaxes(double len=1); //draw axes with real length at the origin
 	virtual void drawnum(double x, double y, int num); //write out the text of a number at the given coordinates.
